// src/proto/generated/id/ac/ui/cs/advprog/kilimanjaro/auth/RequestMetadata.ts
// Original file: public/assets/proto/auth.proto


export interface RequestMetadata {
  'requestId'?: (string);
  'clientVersion'?: (string);
  'correlationId'?: (string);
}

export interface RequestMetadata__Output {
  'requestId': (string);
  'clientVersion': (string);
  'correlationId': (string);
}