// src/proto/generated/id/ac/ui/cs/advprog/kilimanjaro/auth/TokenValidationRequest.ts
// Original file: public/assets/proto/auth.proto

import type { RequestMetadata as _id_ac_ui_cs_advprog_kilimanjaro_auth_RequestMetadata, RequestMetadata__Output as _id_ac_ui_cs_advprog_kilimanjaro_auth_RequestMetadata__Output } from '../../../../../../../id/ac/ui/cs/advprog/kilimanjaro/auth/RequestMetadata';

export interface TokenValidationRequest {
  'metadata'?: (_id_ac_ui_cs_advprog_kilimanjaro_auth_RequestMetadata | null);
  'token'?: (string);
  'includeUserData'?: (boolean);
}

export interface TokenValidationRequest__Output {
  'metadata': (_id_ac_ui_cs_advprog_kilimanjaro_auth_RequestMetadata__Output | null);
  'token': (string);
  'includeUserData': (boolean);
}